Add post searching and filtering
Single posts can be queried, post names can be searched, and posts
belonging to a particular type or containing a particular tag can also
be filtered.

For now, debug information is shown on the front page to make sure
everything is being displayed properly. Theme support will be added
soon, so that posts can actually be displayed as they should be.