Validate SQL queries
Where possible, queries have been validated using regex, with a special
enum type. This includes:

* Slugs (URL friendly version of post titles)
* Tags
* Types
* Dates (Full ISO8601 compliant - eg. 'YYYY-MM-DD HH:mm:ss.ms+HHmm')

Values inserted into the database are also quoted beforehand, to try and
prevent SQL injection. Usually this won't be a problem as the user needs
to be authenticated using their own website first, but of course, this
is a step that *needs* to be taken before ever releasing as stable.

Closes #11