Create post using MicroPub (NOT SECURE)
If you see this commit - don't assume the MicroPub API is done and ready
for you to use on your website. That is far from the case. Currently,
the API is not secured using IndieAuth, and anyone can POST to this URL
and create a post on the database.

Hold your horses!